reworked the directories
